phase 2 extracted
