+++ client/src/components/DashboardPage.jsx
@@ -1,13 +1,12 @@
 // File: client/src/components/DashboardPage.jsx
-// Lines 1-35: Enhanced imports with Phase 1 utility functions
+// Lines 1-25: Enhanced imports with Phase 1 & Phase 2 components
 import React, { useState, useCallback, useMemo } from "react";
 import { useAuth } from "../context/AuthContext";
-import { useNavigate } from "react-router-dom";
+import { useToast } from "../hooks/useToast";
 
 // Custom Hooks - Real API data management
 import useStudentManagement from "../hooks/useStudentManagement";
 import { useDashboardData } from "../hooks/useDashboardData";
-import { useToast } from "../hooks/useToast";
 
 // PHASE 1: Utility Functions - Extracted zero-risk pure functions
 import { 
@@ -17,229 +16,29 @@ import {
   canSendReminder
 } from "../utils/studentCalculations";
 
+// PHASE 2: Extracted Components - UI separation of concerns
+import DashboardHeader from "./dashboard/DashboardHeader";
+import DashboardStatistics from "./dashboard/DashboardStatistics";
+import { LoadingView, ErrorView, ProfileView, EditView } from "./dashboard/DashboardViews";
+
 // Existing Components - Preserved all functionality
-import StatisticsCards from "./dashboard/StatisticsCards";
-import PricingDistribution from "./dashboard/PricingDistribution";
 import StudentManagementSection from "./dashboard/StudentManagementSection";
-import StudentProfileView from "./StudentProfileView";
-import StudentEditForm from "./StudentEditForm";
+import AnnouncementBanner from "./dashboard/AnnouncementBanner";
+
+// Modal Components - Preserved all functionality
 import PaymentModal from "./PaymentModal";
 import AddStudentModal from "./AddStudentModal";
 import SMSCreditsModal from "./modals/SMSCreditsModal";
 import SMSHistoryModal from "./modals/SMSHistoryModal";
 import WeekendEventModal from "./modals/WeekendEventModal";
-import MonthlyReportModal from "./modals/MonthlyReportModal"; // NEW: Monthly Report Modal
-import AnnouncementBanner from "./dashboard/AnnouncementBanner";
+import MonthlyReportModal from "./modals/MonthlyReportModal";
 
-// Lines 35-90: Dark Theme Logout Button Component (PRESERVED)
-const DarkThemeLogoutButton = () => {
-  const { logout } = useAuth();
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    if (window.confirm("Are you sure you want to logout?")) {
-      logout();
-      navigate('/');
-    }
-  };
-
-  return (
-    <button
-      onClick={handleLogout}
-      className="flex items-center justify-center px-4 py-3 bg-red-600 hover:bg-red-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-      title="Logout"
-    >
-      <span className="mr-2">🚪</span>
-      <span>Logout</span>
-    </button>
-  );
-};
-
-// Lines 55-180: ENHANCED DARK THEME HEADER - Mobile-first responsive design with Monthly Report Button (PRESERVED)
-const DarkThemeHeader = ({ 
-  user, 
-  onRefresh, 
-  onOpenCredits, 
-  onOpenHistory, 
-  onOpenWeekendEvent,
-  onOpenMonthlyReport, // NEW: Monthly Report callback
-  loading 
-}) => (
-  <header className="bg-gray-900 border-b border-gray-800">
-    <div className="px-4 py-4 sm:px-6 lg:px-8">
-      {/* Mobile-first layout with proper breathing space */}
-      <div className="flex flex-col space-y-6 lg:flex-row lg:items-center lg:justify-between lg:space-y-0">
-        
-        {/* Title Section - Better mobile typography */}
-        <div className="text-center lg:text-left">
-          <h1 className="text-xl font-bold text-white sm:text-2xl lg:text-3xl">
-            Student Membership Dashboard
-          </h1>
-          <p className="text-gray-400 text-sm mt-2 sm:text-base">
-            Welcome back, {user?.email || 'Administrator'}
-          </p>
-        </div>
-        
-        {/* Action Buttons Section - Mobile-optimized layout */}
-        <div className="flex flex-col space-y-4 sm:space-y-3 lg:flex-row lg:items-center lg:space-y-0 lg:space-x-4">
-          
-          {/* Primary Action Buttons - Enhanced grid layout for mobile */}
-          <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 xl:grid-cols-3 lg:flex lg:items-center lg:gap-3">
-            
-            {/* NEW: Monthly Report Button - Purple theme */}
-            <button
-              onClick={onOpenMonthlyReport}
-              className="flex items-center justify-center px-4 py-3 bg-purple-600 hover:bg-purple-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-              title="Generate Monthly Payment Report"
-              disabled={loading}
-            >
-              <span className="mr-2">📊</span>
-              <span>Monthly Report</span>
-            </button>
-
-            {/* Weekend Event Button */}
-            <button
-              onClick={onOpenWeekendEvent}
-              className="flex items-center justify-center px-4 py-3 bg-orange-600 hover:bg-orange-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-              title="Create Weekend Event"
-              disabled={loading}
-            >
-              <span className="mr-2">📅</span>
-              <span>Weekend Event</span>
-            </button>
-
-            {/* Credits Button */}
-            <button
-              onClick={onOpenCredits}
-              className="flex items-center justify-center px-4 py-3 bg-green-600 hover:bg-green-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-              title="Manage SMS Credits"
-              disabled={loading}
-            >
-              <span className="mr-2">💳</span>
-              <span>Credits</span>
-            </button>
-
-            {/* History Button */}
-            <button
-              onClick={onOpenHistory}
-              className="flex items-center justify-center px-4 py-3 bg-gray-600 hover:bg-gray-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-              title="View SMS History"
-              disabled={loading}
-            >
-              <span className="mr-2">📈</span>
-              <span>History</span>
-            </button>
-            
-            {/* Refresh Button */}
-            <button
-              onClick={onRefresh}
-              disabled={loading}
-              className="flex items-center justify-center px-4 py-3 bg-blue-600 hover:bg-blue-700 text-white rounded-lg transition-colors text-sm font-medium min-h-[44px] disabled:opacity-50 disabled:cursor-not-allowed focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 focus:ring-offset-gray-900"
-              title="Refresh Dashboard Data"
-            >
-              {loading ? (
-                <>
-                  <div className="w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin mr-2"></div>
-                  <span>Loading...</span>
-                </>
-              ) : (
-                <>
-                  <span className="mr-2">🔄</span>
-                  <span>Refresh Data</span>
-                </>
-              )}
-            </button>
-          </div>
-
-          {/* Logout Button Section - Properly separated with visual distinction */}
-          <div className="flex justify-center lg:justify-end pt-3 lg:pt-0 border-t border-gray-700 lg:border-t-0 lg:border-l lg:border-gray-700 lg:pl-4">
-            <DarkThemeLogoutButton />
-          </div>
-        </div>
-      </div>
-    </div>
-  </header>
-);
-
-// Lines 185-260: ENHANCED DARK THEME STATISTICS - Horizontal layout matching design (PRESERVED)
-const DarkThemeStatistics = ({ dashboardData, students, tabCounts, pricingBreakdown }) => {
-  const statisticsData = [
-    {
-      title: "Total Students",
-      value: students?.length || 0,
-      subtitle: "All registered students",
-      icon: "👥",
-      bgColor: "bg-gray-800",
-      iconColor: "text-blue-400"
-    },
-    {
-      title: "Active",
-      value: tabCounts?.active || 0,
-      subtitle: "Currently enrolled",
-      icon: "✅",
-      bgColor: "bg-gray-800",
-      iconColor: "text-green-400"
-    },
-    {
-      title: "Expiring Soon",
-      value: tabCounts?.expiring || 0,
-      subtitle: "Within 7 days",
-      icon: "⚠️",
-      bgColor: "bg-gray-800",
-      iconColor: "text-yellow-400"
-    },
-    {
-      title: "Overdue",
-      value: tabCounts?.overdue || 0,
-      subtitle: "Payment overdue",
-      icon: "🚨",
-      bgColor: "bg-gray-800",
-      iconColor: "text-red-400"
-    },
-    {
-      title: "Monthly Revenue",
-      value: `₱${(pricingBreakdown?.totalRevenue || 0).toLocaleString()}`,
-      subtitle: "Expected monthly",
-      icon: "💰",
-      bgColor: "bg-gray-800",
-      iconColor: "text-green-400"
-    }
-  ];
-
-  return (
-    <div className="mb-8">
-      {/* Statistics Cards - Horizontal layout like your design */}
-      <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-5">
-        {statisticsData.map((stat, index) => (
-          <div 
-            key={index}
-            className={`${stat.bgColor} rounded-lg p-6 border border-gray-700 hover:border-gray-600 transition-colors`}
-          >
-            <div className="flex items-center">
-              <div className="flex-shrink-0">
-                <span className={`text-2xl ${stat.iconColor}`}>{stat.icon}</span>
-              </div>
-              <div className="ml-4 flex-1">
-                <div className="flex items-baseline">
-                  <p className="text-2xl font-semibold text-white">{stat.value}</p>
-                </div>
-                <p className="text-sm font-medium text-gray-300">{stat.title}</p>
-                <p className="text-xs text-gray-500">{stat.subtitle}</p>
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-// Lines 265-315: ENHANCED MAIN DASHBOARD COMPONENT - Complete implementation with imported utilities
+// Lines 35-165: Enhanced Main Dashboard Component - Phase 2 Integration
 export default function DashboardPage() {
   const { user, token } = useAuth();
   const { showSuccess, showError } = useToast();
 
-  // State management - Comprehensive modal and view state (ENHANCED)
+  // State management - Comprehensive modal and view state (PRESERVED)
   const [currentView, setCurrentView] = useState("dashboard");
   const [selectedStudent, setSelectedStudent] = useState(null);
   const [paymentModalOpen, setPaymentModalOpen] = useState(false);
@@ -247,11 +46,11 @@ export default function DashboardPage() {
   const [smsCreditsModalOpen, setSmsCreditsModalOpen] = useState(false);
   const [smsHistoryModalOpen, setSmsHistoryModalOpen] = useState(false);
   const [weekendEventModalOpen, setWeekendEventModalOpen] = useState(false);
-  const [monthlyReportModalOpen, setMonthlyReportModalOpen] = useState(false); // NEW: Monthly Report Modal State
+  const [monthlyReportModalOpen, setMonthlyReportModalOpen] = useState(false);
   const [announcements, setAnnouncements] = useState([]);
   const [smsLoading, setSmsLoading] = useState(false);
 
-  // API data hooks - Enhanced error handling
+  // API data hooks - Enhanced error handling (PRESERVED)
   const { 
     dashboardData, 
     students, 
@@ -261,7 +60,7 @@ export default function DashboardPage() {
     refetch 
   } = useDashboardData(token);
 
-  // Enhanced student management hook - All student operations
+  // Enhanced student management hook - All student operations (PRESERVED)
   const {
     filteredStudents,
     tabCounts,
@@ -283,7 +82,7 @@ export default function DashboardPage() {
     return calculateRevenueData(students);
   }, [students]);
 
-  // NEW: Monthly Report modal handlers - Lines 315-325
+  // Modal handlers - All preserved functionality
   const handleOpenMonthlyReportModal = useCallback(() => {
     setMonthlyReportModalOpen(true);
   }, []);
@@ -374,7 +173,7 @@ export default function DashboardPage() {
       }
 
     } catch (error) {
-      console.error("❌ SMS reminder error:", error);
+      console.error("SMS reminder error:", error);
       
       if (error.message.includes('429')) {
         showError(`Rate limit: Cannot send another reminder to ${student.name} yet. Please wait 24 hours.`);
@@ -456,118 +255,69 @@ export default function DashboardPage() {
     showSuccess('Edit functionality coming soon');
   }, [showSuccess]);
 
-  // Loading state - Enhanced dark theme loading (PRESERVED)
+  // PHASE 2: Loading state using extracted component
   if (loading) {
-    return (
-      <div className="min-h-screen bg-gray-900 flex items-center justify-center p-4">
-        <div className="text-center">
-          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-400 mx-auto mb-4"></div>
-          <p className="text-gray-300 text-lg font-medium">Loading dashboard...</p>
-          <p className="text-gray-500 text-sm mt-2">Fetching student data and statistics</p>
-        </div>
-      </div>
-    );
+    return <LoadingView />;
   }
 
-  // Error state - Enhanced dark theme error handling (PRESERVED)
+  // PHASE 2: Error state using extracted component
   if (error) {
-    return (
-      <div className="min-h-screen bg-gray-900 flex items-center justify-center p-4">
-        <div className="bg-gray-800 rounded-lg shadow-lg p-6 max-w-md w-full text-center border border-gray-700">
-          <div className="text-red-400 mb-4 text-4xl">❌</div>
-          <h3 className="text-xl font-semibold text-white mb-2">Dashboard Error</h3>
-          <p className="text-red-400 text-sm mb-6">{error}</p>
-          <button 
-            onClick={refetch} 
-            className="w-full px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors font-medium"
-          >
-            Retry Loading
-          </button>
-        </div>
-      </div>
-    );
+    return <ErrorView error={error} onRetry={refetch} />;
   }
 
-  // Profile view - Enhanced dark theme profile (ENHANCED with Monthly Report)
+  // PHASE 2: Profile view using extracted component
   if (currentView === "profile" && selectedStudent) {
     return (
-      <div className="min-h-screen bg-gray-900">
-        <DarkThemeHeader 
-          user={user}
-          onRefresh={refetch}
-          onOpenCredits={() => setSmsCreditsModalOpen(true)}
-          onOpenHistory={() => setSmsHistoryModalOpen(true)}
-          onOpenWeekendEvent={handleOpenWeekendEventModal}
-          onOpenMonthlyReport={handleOpenMonthlyReportModal} // NEW: Pass Monthly Report handler
-          loading={loading}
-        />
-        <main className="px-4 py-6 sm:px-6 lg:px-8">
-          <div className="max-w-4xl mx-auto">
-            <button
-              onClick={handleBackToDashboard}
-              className="mb-6 flex items-center text-blue-400 hover:text-blue-300 text-sm font-medium min-h-[44px] transition-colors"
-            >
-              ← Back to Dashboard
-            </button>
-            <StudentProfileView
-              student={selectedStudent}
-              onBack={handleBackToDashboard}
-              onEdit={handleEditStudent}
-            />
-          </div>
-        </main>
-      </div>
+      <ProfileView
+        user={user}
+        selectedStudent={selectedStudent}
+        onBack={handleBackToDashboard}
+        onEdit={handleEditStudent}
+        onRefresh={refetch}
+        onOpenCredits={() => setSmsCreditsModalOpen(true)}
+        onOpenHistory={() => setSmsHistoryModalOpen(true)}
+        onOpenWeekendEvent={handleOpenWeekendEventModal}
+        onOpenMonthlyReport={handleOpenMonthlyReportModal}
+        loading={loading}
+      />
     );
   }
 
-  // Edit view - Enhanced dark theme editing (ENHANCED with Monthly Report)
+  // PHASE 2: Edit view using extracted component
   if (currentView === "edit" && selectedStudent) {
     return (
-      <div className="min-h-screen bg-gray-900">
-        <DarkThemeHeader 
-          user={user}
-          onRefresh={refetch}
-          onOpenCredits={() => setSmsCreditsModalOpen(true)}
-          onOpenHistory={() => setSmsHistoryModalOpen(true)}
-          onOpenWeekendEvent={handleOpenWeekendEventModal}
-          onOpenMonthlyReport={handleOpenMonthlyReportModal} // NEW: Pass Monthly Report handler
-          loading={loading}
-        />
-        <main className="px-4 py-6 sm:px-6 lg:px-8">
-          <div className="max-w-2xl mx-auto">
-            <button
-              onClick={() => setCurrentView("profile")}
-              className="mb-6 flex items-center text-blue-400 hover:text-blue-300 text-sm font-medium min-h-[44px] transition-colors"
-            >
-              ← Back to Profile
-            </button>
-            <StudentEditForm
-              student={selectedStudent}
-              onBack={handleBackToDashboard}
-              onSave={handleEditSave}
-            />
-          </div>
-        </main>
-      </div>
+      <EditView
+        user={user}
+        selectedStudent={selectedStudent}
+        onBack={() => setCurrentView("profile")}
+        onSave={handleEditSave}
+        onRefresh={refetch}
+        onOpenCredits={() => setSmsCreditsModalOpen(true)}
+        onOpenHistory={() => setSmsHistoryModalOpen(true)}
+        onOpenWeekendEvent={handleOpenWeekendEventModal}
+        onOpenMonthlyReport={handleOpenMonthlyReportModal}
+        loading={loading}
+      />
     );
   }
 
-  // Main dashboard view - Enhanced dark theme implementation (ENHANCED with Monthly Report)
+  // PHASE 2: Main dashboard view using extracted components
   return (
     <div className="min-h-screen bg-gray-900">
-      <DarkThemeHeader 
+      {/* PHASE 2: Extracted Header Component */}
+      <DashboardHeader 
         user={user}
         onRefresh={refetch}
         onOpenCredits={() => setSmsCreditsModalOpen(true)}
         onOpenHistory={() => setSmsHistoryModalOpen(true)}
         onOpenWeekendEvent={handleOpenWeekendEventModal}
-        onOpenMonthlyReport={handleOpenMonthlyReportModal} // NEW: Pass Monthly Report handler
+        onOpenMonthlyReport={handleOpenMonthlyReportModal}
         loading={loading}
       />
 
       <main className="px-4 py-6 sm:px-6 lg:px-8">
         <div className="max-w-7xl mx-auto space-y-8">
-          {/* Announcements Banner */}
+          {/* Announcements Banner - PRESERVED */}
           {announcements.length > 0 && (
             <AnnouncementBanner 
               announcements={announcements}
@@ -576,15 +326,15 @@ export default function DashboardPage() {
             />
           )}
 
-          {/* Statistics Section - Enhanced dark theme */}
-          <DarkThemeStatistics 
+          {/* PHASE 2: Extracted Statistics Component */}
+          <DashboardStatistics 
             dashboardData={dashboardData}
             students={students}
             tabCounts={tabCounts}
             pricingBreakdown={revenueData}
           />
 
-          {/* Student Management Section - Enhanced container */}
+          {/* Student Management Section - PRESERVED with enhanced container */}
           <div className="bg-gray-800 rounded-lg border border-gray-700 overflow-hidden shadow-lg">
             <StudentManagementSection
               filteredStudents={filteredStudents}
